Separated common component into opencv_loader.h
